+++ src/lib/mrvFLTK/mrvCallbacks.cpp
@@ -130,9 +130,6 @@ namespace mrv
         {_("Histogram"), (Fl_Callback*)histogram_panel_cb},
         {_("Logs"), (Fl_Callback*)logs_panel_cb},
         {_("Media Information"), (Fl_Callback*)image_info_panel_cb},
-#ifdef MRV2_NETWORK
-        {_("Network"), (Fl_Callback*)network_panel_cb},
-#endif
 #ifdef TLRENDER_NDI
         {_("NDI"), (Fl_Callback*)ndi_panel_cb},
 #endif
@@ -1177,10 +1174,6 @@ namespace mrv
         if (panel::ndiPanel)
             panel::ndiPanel->save();
 #endif
-#ifdef MRV2_NETWORK
-        if (panel::networkPanel)
-            panel::networkPanel->save();
-#endif
 #ifdef TLRENDER_USD
         if (panel::usdPanel)
             panel::usdPanel->save();
@@ -1217,10 +1210,6 @@ namespace mrv
         panel::filesPanel = nullptr;
         delete panel::comparePanel;
         panel::comparePanel = nullptr;
-#ifdef MRV2_NETWORK
-        delete panel::networkPanel;
-        panel::networkPanel = nullptr;
-#endif
 #ifdef TLRENDER_NDI
         delete panel::ndiPanel;
         panel::ndiPanel = nullptr;