+++ src/lib/mrvNetwork/mrvCommandInterpreter.cpp
@@ -1282,18 +1282,6 @@ namespace mrv
                 if ((!value && webrtcPanel) || (value && !webrtcPanel))
                     webrtc_panel_cb(nullptr, ui);
             }
-            else if (c == "Network Panel")
-            {
-                bool receive = prefs->ReceiveUI->value();
-                if (!receive)
-                {
-                    tcp->unlock();
-                    return;
-                }
-                bool value = message["value"];
-                if ((!value && networkPanel) || (value && !networkPanel))
-                    network_panel_cb(nullptr, ui);
-            }
             else if (c == "USD Panel")
             {
 #ifdef TLRENDER_USD