+++ src/lib/mrvWidgets/mrvVersion.h
@@ -23,6 +23,8 @@ namespace mrv
     const std::string build_date();
     const std::string build_info();
     const std::string running_info();
+    const std::string cpu_info();
+    const std::string gpu_info();
 
     void about_message(mrv::TextBrowser* b);
     void thanks_message(mrv::TextBrowser* b);